+++ packages/cli/src/__tests__/e2e/p1/create.spec.ts
@@ -17,7 +17,8 @@ Options:
 
 Commands:
   wasm [options] <language> <name>    Create a Polywrap wasm wrapper. langs:
-                                      assemblyscript, rust, golang, interface
+                                      assemblyscript, rust, golang, typescript,
+                                      interface
   app [options] <language> <name>     Create a Polywrap application. langs:
                                       typescript, python, rust, android, ios
   plugin [options] <language> <name>  Create a Polywrap plugin. langs: