+++ packages/cli/src/lib/project/manifests/language.ts
@@ -8,3 +8,7 @@ export type AnyProjectManifestLanguage =
   | PolywrapManifestLanguage
   | PluginManifestLanguage
   | AppManifestLanguage;
+
+export type BuildableLanguage =
+  | Exclude<PolywrapManifestLanguage, "interface" | "wasm/typescript">
+  | "wasm/javascript";