+++ packages/schema/bind/src/bindings/golang/transforms/index.ts
@@ -1,2 +1,3 @@
+export * from "./appendImportedTypes";
+export * from "./extractImportedTypes";
 export * from "./moduleNeedsTypes";
-export * from "./moduleNeedsImportedTypes";