+++ packages/templates/wasm/typescript/src/index.ts
@@ -1,0 +1,7 @@
+import { Args_foo, ModuleBase } from "./wrap";
+
+export class Module extends ModuleBase {
+  foo(args: Args_foo): string {
+    throw new Error("Not implemented");
+  }
+}